pyproject.toml
[tool.poetry]
name = "diplom-work-education-modules-"
version = "0.1.0"
description = ""
authors = ["berestw <paladin.berestw@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
psycopg2-binary = "^2.9.9"
black = "^24.8.0"
ipython = "^8.26.0"
getenv = "^0.2.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
setuptools = "^72.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
